requirements.txt
gunicorn
Pygments
Markdown
djongo
jieba
pika
mysqlclient
lxml
urllib3
pillow
requests
httplib2
six

django == 2.1.4
djangorestframework
django-reversion
django-cors-headers
django-filter
django-crispy-forms
django-formtools
django-cors-headers
django-picklefield
django-mptt
django-import-export
django-advanced-filters
django-bulk-update
django-ckeditor==5.6.1
